+++ src/components/mod.rs
@@ -9,3 +9,4 @@ pub mod remote;
 pub mod score_card;
 pub mod settings;
 pub mod shortcuts;
+pub mod up_next;